1.专题界面搭建 2.专题界面数据显示，添加了刷新框架

+++ app/src/main/java/com/anfly/anflyshop/model/bean/SpecialTopicBean.java
@@ -103,7 +103,7 @@ public class SpecialTopicBean {
 
             private int id;
             private String title;
-            private int price_info;
+            private float price_info;
             private String scene_pic_url;
             private String subtitle;
 
@@ -123,11 +123,11 @@ public class SpecialTopicBean {
                 this.title = title;
             }
 
-            public int getPrice_info() {
+            public float getPrice_info() {
                 return price_info;
             }
 
-            public void setPrice_info(int price_info) {
+            public void setPrice_info(float price_info) {
                 this.price_info = price_info;
             }
